Do not compute motif area if image is not yet ready
Trying to compute the position of the motif area element, when image
width and height have not yet been determined leads to `NaN`
values. Since `NaN != NaN`, the `onUpdate` callback is then called on
every render leading to an infinite loop.

This causes freezes the editor soon as an uploading or unprocessed
image is selected as backdrop of a section.

REDMINE-17839