Consolidate knowledge about event context data in one place
Make `MediaPlayer` just pass along event context data to `getPlayer`.
Instead already wrap the event object placed on the context in a
`page` property. That way all information about the shape of event
data can be found in `useEventContextData.js`.

Since `MediaPlayer` no longer cares about the shape of the event
context data, we can change the spec to just place some dummy data on
the context and expect it to be passed to `getPlayer`. We deliberately
use some obviously made up piece of data instead of trying to
replicate the shape of the real data here which will be subject to
change in the future.

We also fix a naming inconsistency by also calling the prop
`mediaEventsContextData` instead of just `mediaEventsContext` in
`PlayerContainer`.

REDMINE-17754