Use correct section index for EventContextDataProvider
We need to pass the index of the rendered section, not the current
section. Passing the current section index causes the wrong section to
be reported in media events when a video is played that is not in the
current section but still visible in the viewport.

It also causes extra context updates and rendering.

REDMINE-17754