Turn react and react-dom into peer dependencies of pageflow-scrolled
Since only one copy of React is allowed to be loaded at a time, soon
as other Paegflow plugin packages that define content elements are
added to a host application, we need to make sure that all Pageflow
packages share a React version.

`pageflow-scrolled` can therefore no longer declare `react` as a
normal dependency, but instead needs to rely on the host application
to install a shared copy of React.

When running `yarn install` in the root of the `pageflow` repository,
React packages will still be installed into the root `node_modules`
directory since they are transitive dependencies of other packages
`pageflow-scrolled` depends on.

This has the positive side effect that - for example when running Jest
- React can still be imported. On the other hand, when linking the
`pagefow-scrolled` package in a host application, React will now be
loaded from that root `node_modules` directory, while other (not
linked) packages in the host application still load React from the
host application's `node_modules` directory. To prevent errors, we
therefore need to link the `react` package as described in the updated
"Node pacakge development" guide.

REDMINE-17383