Server side render scrolled entries
Make sure HTML content of the entry is present in the server rendered
response. Before, all visible content was only rendered client side
based on the JSON seed embedded in the document. Now we can at least
display the textual content even if the user has JS turned off. Images
and media will not be visible since lazy loading still depends on
JS. This is also an improvement in regard to SEO since scrapers will
not need to run JS to obtain the content.

REDMINE-17566