Accept entry templates as parameter to config_for
To have the necessary data available on entry template even when
calling `new`, we need to add account and entry template to it from
the params.

Since config_for expects an entity that responds to
`enabled_feature_names`, we delegate that one to an entry template's
account.

Finally, we dynamically add theme select options and widget fields to
entry template form. For theme validation to not allow themes of other
entry types as well, we adjust `available_themes`. Also, to avoid
listing widget types in the editor that aren't available for the
entry's type, we adjust `resolve_widgets`.

Also includes some renamings from entry template's entry_type to
entry_type_name.

REDMINE-17793