Render structured data in Video Player
Include structured data for inline and backdrop videos.

`durationInMs` is already included in the server rendered JSON seed
used for published entries [1, 2]. We need to add a mapping to
`watchCollections` to make sure it is also available in the editor
where entry seed data is obtained from Backbone models.

REDMINE-17390

[1] https://github.com/codevise/pageflow/blob/5f59d03471416b6e19dab2cdfcf8bf0649520c3b/app/views/pageflow/video_files/_video_file.json.jbuilder#L1

[2] https://github.com/codevise/pageflow/blob/5f59d03471416b6e19dab2cdfcf8bf0649520c3b/entry_types/scrolled/app/views/pageflow_scrolled/entry_json_seed/_entry.json.jbuilder#L72