Ensure media has loaded before generating waveform
The previous implementation checked to see if the audio element still
references blank sources in the Waveform render function and delayed
rendering the Wavesurfer component.

It needed to use DOM API to look up the media element during
render. This is an anti pattern and lead to exceptions when React
updated the DOM after evaluating the render function so that the media
element could still be found during render, but was already gone when
the Wavesurfer component mounted.

The core issue that is being worked around here is that calling
VideoJS's `src` to update the sources of a media element does not take
effect immediately. Instead VideoJS wraps the update in a `setTimeout`
[1]. This causes the sources not yet to have been updated once a
player is requested and appended to the DOM.

As an alternative solution, we now wait for an initial
`loadedmetadata` event in `watchPlayer` and only save the media
element id once it is dispatched. This ensures that the media element
has already been updated with the correct sources when the `Waveform`
component is rendered with a `mediaElementId`.

REDMINE-17677

[1] https://github.com/codevise/pageflow/blob/610793b9a76e2795bf544cdb9cf35b4035dec3b8/vendor/assets/javascripts/videojs.js#L11144