Hide mobile sharing icons on desktop
Mobile sharing icons don't need to be
displayed in a desktop version.

Since editor has an emulation mode which
won't work with a normal mobile device identifying
hook. It creates separate views for editor and published view

In view it checks for the device and
simply hides the sharing icons.

In editor it listens for a post message
on emulation mode event change and hides
the icons accordingly