Prevent empty space below entry
Reveal/fade out sections need to make their backdrop longer to apply
the clipping effect only at the top. Using height 200% works is the
same as 200vh as long as the section does not have a lot of
content. For longer sections this causes the backdrop to extend beyond
the next section. Change height to account for longer sections.

REDMINE-17838