Add useBrowserFeature to fix usePhonePlatform in SSR
Browser feature detection is not available during server side
rendering. A first idea might be to simply let all feature flags
default to `false` on the server. We need to be careful, though, how
we implement this to prevent mismatches when server side rendered
markup is hydrated. In particular, the first render on the client
needs to generate the same result as the server side rendered markup.

We therefore introduce a new hook `useBrowserFeature` which will
return false in SSR, but also during the first render. Once components
are being mounted, it triggers a re-rerender with the actual feature
states.

To make the `usePhonePlatform` hook with its new dependencies work in
tests, we move both the `BrowserFeaturesProvider` and the
`PhonePlatformProvider` to the `RootProviders` and use the
`renderHookInEntry` test helper instead. This also ensures that
`usePhonePlatform` can now be used everywhere, not just in the
`AppHeader`.

Since `browser` is now uses by the `RootProviders` we import the
`fakeBrowserFeatures` test helper globally, to make sure feature
detection has finished when components are being rendered just like in
production code.