Do not fail if media element for waveform cannot be found
When adding a caption to a waveform audio content element, the
`Waveform` and the `MediaPlayer` components get unmounted and
recreated wrapped in a `Figure` component. Removing the `MediaPlayer`
dispatches the `discardMediaElementId` action, but this only gets
processed for the next render. When the `Waveform` component is
mounted wrapped in the `Figure` component, the stored media element id
has not yet changed, still the `MediaPlayer` component is not yet
mounted again. Trying to look up the media element fails.

If we ignore the missing media element. The `Waveform` will get
unmounted again once the `discardMediaElementId` action has been
processed. Once the new `Audio` component has been mounted it
dispatches a `saveMediaElementId` action which in turn causes a new
`Waveform` component to be rendered. This time looking up the media
element succeeds.

REDMINE-17677