Make motif area selection snap to grid
Once background position depends on motif area, it will be important
to assign identical motif areas to backdrop images of adjacent
sections when using section transitions to build scroll animations.

Making the selection rect snap to a grid makes it more likely to
select exactly the same motif area when drawing roughly the same
rectangle.

Grid size is calculated such that the longer side of the image is
always divided into 20 cells. For the other dimension, cell size is
increased just enough to ensure the length of the shorter side of the
image is a multiple of the cell size.

Hide overflow of preview thumbnail to prevent rounding errors from
making mask element extend beyond image by a pixel.

REDMINE-17967