Make it easier to use hooks in renderInEntry tests
When rendering components that expect the result of certain hooks as a
prop, the hook could previously not be called directly in the test
since hooks can only be invoked in the context of a component.

One workaround was to first call `renderHookInEntry` with the same
seed and pass `result.current` to the component rendered via
`renderInEntry`. This is a lot of boilerplate.

To make it easier to test that a component works correctly with the
result of hook, instead of a React component, you can now pass a
function which returns a React component to `renderInEntry`. The
function is evaluated inside an one-off helper component which makes
it possible to use hooks.

REDMINE-17879