Add toggle mute button to navigation
Clearly indicate to the reader whether sound is on or off and make it
easy to toggle. Move logo in phone layout to the left to prevent
collision with button.

As requested by customer always color icon blue to make it stand out.

To make the `useMediaMuted` hook work in `AppHeader`, the
corresponding context needs to be pulled up the the
`RootProviders`. This requires adapting `MediaPlayerControls` specs,
which before were not rendered inside `MediaMutedProvider`. Now media
is recognized as muted, which causes captions to be enabled by
default. This leads to a different label for the auto text track menu
item. Turning the example text track into a subtitle, decouples the
test from this special behavior.

Use buttons instead of href-less anchors for navigation buttons to
make them keyboard accessible. Use flex instead of float right to make
the visible and tab order match the DOM order.

REDMINE-17965