Make players take poster file instead of poster id
Pulling file lookup up into the content element lets us use the poster
file instead of the id to determine whether audio controls are in
stand alone mode. This ensures the mode is set correctly, when an
image file is deleted from the entry: The id might still be present in
the configuration, but the file itself can no longer be looked up.

Already before, we started moving towards passing file objects instead
of ids for the actual audio and video files. So this change makes
things more consistent.

REDMINE-17746