Unmute every time play button is clicked
So far `playBlessed` action only unmuted the first time it is
dispatched. This made it impossible to pause players by clicking the
poster since clicking the poster always tried to unmute by dispatching
`playBlessed`. Before the introduction of the mute/unmute button, this
could not happen since there was no way to mute the entry again.

REDMINE-17746