Add drag handle support to selection rect
Allow passing a drag ref to selection rect and render move icon in the
top let corner. For full width selections the move icon is positioned
on top of the content element. in this case it is always white with a
dark shadow to ensure visibility on light backgrounds.

REDMINE-17331