Render custom drop targets for text blocks
To allow moving content elements between paragraphs (or other
children) of a text block, we need to render custom drop targets. Soon
as an element is dragged over a text block, we measure the children
and cover the element with divs for each "gap" beween two text block
children. That way dragging an element to the lower half of child or
the upper half of the next child results in the same drop indicator
being displayed between the two children.

REDMINE-17331