Allow moving content elements via batch request
The concepts of splitting and merging content elements were already
present in the functions to insert and delete content elements.
Content elements can be inserted at split points of other content
elements (e.g. between paragraphs of text blocks). Deleting content
elements supports merging the newly adjacent content elements
(e.g. two text blocks surrounding a deleted image).

When moving content elements, the same concepts are at play but in
more complicated combinations: When moving an image from in between
two text blocks into a another text block or even to a split position
in one of the adjacent text blocks, both merge and split operations
need to be performed in the context of a single batch update.

The existing `batch` function which was used to construct batch
requests relied on data read from a section's actual content elements
collection and allowed manipulating the resulting batch items to
perform advanced transformations. In particular, it manually rolled
back changes after a failed request. For the more involved cases
needed now this became overly complex and hard to think about.

We therefore introduce a new `Batch` class which allows recording a
set of changes that shall be performed on the content elements of a
section via an imperative interface. The changes are performed on a
shallow copy of the section's data, which makes creating the required
batch items easy. Changes are applied to the actual models and
collections only after the request succeeds. Hence, there is no longer
a need to think about rolling back changes.

The following commit rewrites the insert and delete functions to use
the same infrastructure rendering the old batch functionality
obsolete.

All existing tests in this area already only observed the triggered
requests and final model changes. This makes the mechanism to
construct batch requests an implementation detail. As a nice side
effect, we will be able to reuse most of the insert/delete tests after
the rewrite. The resulting tests are rather long and verbose, but
capture the things we actually care about. We therefore continue to
test the new batch functionality using this black box approach.

REDMINE-17331