Add ThirdPartyConsent component and context provider
OptIn and OptOut are decorator components that are applied by
ThirdPartyConsent.

== Opt-in decorator component ==

OptIn receives a consent method that will take care of several
things:

- call giveConsent with provider name as argument. giveConsent is
taken from third party consent context and will set the consent in the
cookie as well as in third party consent context,
- call setConsentedHere, which serves as a marker to distinguish on
which embed an opt-in was given. This marker will be passed through
ThirdPartyConsent and be passed to children if children is a render
prop, i.e., a function. In the example of video embeds, if its value
is true, it will lead to the actual video being shown instead of a
poster image, but only for the one video where the user consented
- call activate embed, which can be given as a callback to
ThirdPartyConsent and could e.g. be used to play video without the
user needing to click another time

OptIn also receives a cookie message that prompts the user to consent,
and a height for the opt-in prompt.

The fill color that is used for icon styles is temporarily taken from
a customer theme. Themes are still WIP.

== Opt-Out decorator component ==

The opt-out component just adds an icon button that displays a
tooltip on focus or hover. The tooltip contains a link that leads
e.g. to a page where consent can be managed. Probably soon, we'll add
ways to manage consent within Pageflow.

hideTooltip prop can be used for cases like a playing video, when the
tooltip icon could be a nuisance.

optOutPlacement prop can set a CSS class that controls placement of
the tooltip icon, because different types of embeds have different
content layouts.

== ThirdPartyConsent component ==

Basically, this component decides whether children should be
decorated, and if so, whether either with an opt-in prompt or with an
opt-out tooltip icon.

children could be a render prop, i.e. a function.

providerName is used as a key to look up provider-specific
consent.

noop determines whether ThirdPartyConsent should just display children
undecorated.

All other props are passed through to OptIn and OptOut decorator
components.

Display of opt-in and opt-out are conditional on theme options: a
cookie name for the opt-in, and an opt-out link for the opt-out.

In this iteration, the editor and the preview thumbnails won't display
anything that differs from the state before this change.

== ThirdPartyConsentProvider ==

ThirdPartyConsentProvider puts two things onto a context: An object
containing consents encoded as boolean values, and a giveConsent
function that can be used to give consent to use content from a third
party provider.

getProviderCookieKey/getProviderPageflowKey are used to translate keys
for the case that some customer would like to reuse their own cookie
and has keys that aren't named like the Pageflow defaults -- in our
iteration, 'video', and 'datawrapper'.

getCookieContent will deserialize cookie content and use Pageflow's
default keys in the object it returns. setCookieContent will
serialize, using customer-specific keys where needed.

giveConsent is put on the context. Calling it with a provider name
will update the cookie as well as the context.

data-save logo is used for the opt-in prompt.

REDMINE-17863