Add globals for pageflow package to shared Jest config
Files like `editor/models/EncodedFile.js` depend on the
`pageflow.config` global via the `$state` alias. In the test suite of
the `pageflow` package itself, so far this was worked around by
mapping the `$state` alias to a custom support file that exports an
object of the desired shape.

When we want to use `EncodedFile` (or another module depending on it)
in the specs of a package like `pageflow-scrolled`, we cannot re-map
the alias. Since `$state` already relies on the `pageflow` global if
present, we instead provide a shared piece of Jest configuration to
define the `pageflow` global. This can then be used in both the
`pageflow` and the `pageflow-scrolled` package. That way knowledge
about the exact shape of the `pageflow` global remains inside the
`pageflow` package.