Prevent wrapping placeholder of heading
Now that heading use word wrap, the german heading placeholder
"Überschift" also is too long and wraps into two lines. This only
happens on the first page, when the heading is rendered in a larger
font since it is supposed to be the entry title.

Change placeholder to say "Title"/"Titel" in this case, which is
shorter and also better expresses the role of the element on the first
page.

REDMINE-18044