Fix intersection observing when moving elements in the DOM
When ordering elements, adjacent content elements sometimes become
unprepared. This was caused by `useOnScreen` reporting that the
content element is no longer intersecting the viewport.

Turns out Chrome invokes the intersection observer callback with
multiple entries when elements are moved in the DOM. Before, we only
expected a single entry to be reported to the callback. We now use the
last entry assuming that it contains the most up to date information
regarding intersection state.