Improve SSR compatiblity of ThirdPartyConsentProvider
Read cookie in effect and store its value in state. This ensures that
consent state will be empty in initial render (just like in SSR) when
the DOM is hydrated. This makes it possible to use components that
depend on consent state in SSR without creating mismatches between
client and server rendered DOM.

So far we relied on content elements to only render consent components
once lazy loading had been triggered for the content element. This
sidestepped the problem since lazy loading never happens in SSR. With
this change, we are free to use consent components where ever we
please in content elements.

REDMINE-17863