Adjust styles for configurable Datawrapper backgrounds
Charts have since gained an option to make background color of the
content element configurable to accommodate charts with transparent
backgrounds. Also a padding has been added to the content element that
is now visible around the opt in component using the configured
background color.

We therefore change the styles to use the current color - displaying
the opt in on a black backgrund for video embeds. For charts we
calculate an appropriate text color to ensure contrast with the
configured background.

REDMINE-17863