Don't set consent cookie if domain does not match
Mobile Safari ignores the cookie. Chrome sets falls back to the
current hostname. When viewing an entry in the preview, we want to
prevent setting the cookie for the current domain, since it does not
match the domain of the opt out url and hence can not be unset again.

With this change, embeds now always need to be enabled in the
preview. This might be too annoying, but could also be seen as a
reminder for the editor that first time visitors will not see these
embeds immediately.

If we want to set the cookie again, we will also need some way to opt
out in preview mode.

REDMINE-17863