Improve third party opt in render function
Pass object with `consentedHere` property instead of the boolean
itself to render function. This makes the data passed to the render
function more expressive. Otherwise we have to rely on the component
implementing the render function to give the positional parameter the
correct name to preserve meaning. Passing an object also will let us
add additional properties if there should ever be more info to pass
along.

REDMINE-17863