Make shadow opacity configurable
Add separate options for static shadow that is either displayed as a
gradient for left/right layout or translucent scrim for centered
layout and dynamic shadow that is faded in when motif area intersects
with content. Use static shadow opacity for vertical gradient that is
displayed behind content when content is padded.

REDMINE-18119