Do not fall back to file default motif area
When an image file is used as both desktop and mobile image and a
motif area is selected for the desktop variant, the mobile variant
will start using the default motif area stored in the file
configuration.

We accept the breaking change since there are no important old entries
yet.

REDMINE-18038