Fix scroll flickering on Android Chrome
For unknown reasons, if `touch-action: none` is set on the navigation
bar, tapping somewhere inside the navigation bar, causes weird
rendering glitches when scrolling the page from then on. Especially
when scrolling the page via a flick gesture, once the scroll animation
stops, the page is sometims briefly rendered as if the viewport height
had changed: a white rect appears at the bottom of the page and
covering background images are briefly moved to an incorrect position.

The same effect is visible if scrolling is blocked by preventing
default on `touchmove` instead of applying `touch-action: none`. To
still prevent the mobile navigation from scrolling the page, we force
it to have scroll overflow and use `overscroll-behavior`.

REDMINE-18162