Invisible share links should not be clickable
So far `pointer-events: all` was used to re-enable pointer events for
an element inside a `pointer-events: none` parent. The correct way to
reset pointer events is to use `auto`, though. `all` is reserved for
SVG and causes SVG elements to receive pointer events even if
`visibility` is `hidden`. This caused share icons in the hidden
tooltip to be clickable.

Replace all occurences across all CSS files.

REDMINE-18174

[1] https://developer.mozilla.org/de/docs/Web/CSS/pointer-events