Ensure muted inline media pauses on deactivation
When scrolling away from a media element that is playing without sound
while the audio context has not yet been resumed (i.e. auto play is
still forbidden by the browser), the controls switch to paused state,
but the element keeps playing. This can be observed in Chrome before
the first user interaction with the document.

The problem is caused by the fact that `AudioContext#resume` returns a
promise that stays pending when auto play is forbidden.
`tryResumeIfSuspended` [1] in the `volumeFading/webAudio` module
therefore also stays pending. This causes `fadeVolume` to not
resolve. `fadeOutAndPause` [2], which is called on deactivation [3],
thus does not pause in this case. At least in Scrolled, players will
always be muted in this situation. In the general case it is not
clear, though, whether for a muted player `fadeOutAndPause` should
still wait for the delay to pass before resolving. Also in Paged,
`fadeOutAndPause` can be called on `audio5js` players which do not
implemented a `muted` method.

For the Scrolled MediaPlayer, it makes sense to immediately pause on
`fadeOutAndPause` if the player is muted. So this looks like a safe
place to work around the problem.

REDMINE-18168

[1] https://github.com/codevise/pageflow/blob/62b7c1a8304c97d8fef9d5cdcfdeb143ed6159e1/package/src/frontend/mediaPlayer/volumeFading/webAudio.js#L23
[2] https://github.com/codevise/pageflow/blob/62b7c1a8304c97d8fef9d5cdcfdeb143ed6159e1/package/src/frontend/mediaPlayer/volumeBinding.js#L48
[3] https://github.com/codevise/pageflow/blob/62b7c1a8304c97d8fef9d5cdcfdeb143ed6159e1/entry_types/scrolled/package/src/contentElements/inlineVideo/InlineVideo.js#L31