Fix classic loading spinner preview in editor
Before the paged entry type was extracted, the editor used the same
layout as the published entries. Once the paged entries used their own
layout [1], the editor got a separate layout [2] and everything
related to Paged was removed [3]. We missed that the loading spinner
inline script was also needed inside the editor to ensure the preview
could be displayed. Add it back to the paged head editor partial.

We also could have added the rule to an editor stylesheet, but that
would have meant introducing a third place where loading spinner
display is controlled. This way editor and published entry work the
same way.

REDMINE-17738

[1] https://github.com/codevise/pageflow/commit/b0762674c999140021dc725f4445d21b3e72a417#diff-d6c740536eb9bfa6732ef7dfa73df647cae865b6d4a0f7d48ea309ded4c745aa
[2] https://github.com/codevise/pageflow/commit/1ee34a99ca0ae79aa6d700a19a21b3b26839ded2#diff-bf899f40e327b5be0b920c45d6a8602eddfbd0bdbf90e1cdd312c696cec13ec9
[3] https://github.com/codevise/pageflow/commit/e83409db124ead25864ce9def998e781809b0065#diff-bf899f40e327b5be0b920c45d6a8602eddfbd0bdbf90e1cdd312c696cec13ec9