Prevent failing section order request when deleting chapter
Chapter sections collection is cleared when chapter is destroyed. The
`remove` event triggered consolidation of positions and an order
update request. The request failed since the chapter was already
destroyed on the server.  We do not care about gaps in section
`position` attributes and can thus turn off consolidation.

REDMINE-18272