Improve before/after wiggle keyframe animation
Explicitly define 0% and 100% frames to prevent animation errors of
clipped region in Safari. `animation-fill-mode` is no longer needed.

Use `margin-left` instead of `transform` to animate slider. `clip`
appears to not be handled by the compositor. Combining it with a
compositor supported property like `transform` causes the animations
to get out of sync in Safari.

REDMINE-18076