Do not trigger multiple save requests for backdrop updates
Setting section configuration attributes like `backdropType` or
`backdropImage` also updates the `backdrop` attribute which is read in
frontend code. So far these updates each triggered a save request,
which could lead to race conditions when different Passenger processes
handled these requests concurrently since the first request still
contained an old value for the `backdrop` attribute.

Perform both updates in a single `set` call. It would probably be
better to not store `backdrop` at all and compute it in the structure
hooks. For now, we went with the smaller change, though.

REDMINE-18462