Add position wide for heading content element
To prevent long words in headings from being wrapped and hyphenated,
we want to allow greater line width for heading content
elements. Especially for headings in the first section, a large font
is used. With sections' default max width, only very short words fit
into a single line.

For sections using the "cards" appearance variant, we still want to
prevent headings from extending beyond the edge of the card. We
considered different options:

* Making the card wider if it contains a wide heading.

  This leads to complicated follow-up problems regarding the interplay
  with sticky elements:

  Since there is less space next to the card, sticky elements have to
  be inlined earlier, when viewport width decreases.  Especially when
  using the editor on a smaller screen, this can lead to cases where a
  sticky position setting never takes effect during editing.
  Inconsistent configurations can then go unnoticed until the entry is
  viewed on a larger screen.

  Moreover, using the shadow appearance, a wide heading followed by a
  sticky image does not risk visual collision and could be seen as a
  valid use case. This leads to the question whether interplay of
  sticky and wide elements should then also depend on the chosen
  appearance somehow.

  Finally, if cards can be made wider by placing a wide heading
  inside, it seems inconsequential to not also let the other content
  elements of the section use the full card width. This then leads to
  the idea to make section default line width configurable, which is
  currently coupled to section layout (left/right: narrow, center:
  wider). This might be a desirable feature in itself, but also makes
  the section settings form more complex. We currently do not have
  sufficient information about actual customer needs to confidently
  make this trade-off.

  Due to these open questions, we abandoned this option.

* Placing the heading outside of the card

  Already before, sticky and full elements were placed outside of
  cards. We can think of wide elements as being placed in sections
  just like full elements, except for still having left and right
  margin. Making wide element behave like full elements also results
  in a simple solution regarding sticky elements: Wide elements never
  move past sticky elements, but instead push sticky elements out of
  the viewport. The card width stays fixed and can thus not conflict
  with sticky images.

  It could be seen as a downside of this approach, that it results in
  two ways to achieve a visually similar outcome: Placing the header
  in a different section that does not use card appearance but has the
  same backdrop. Since the main requirement formulated by the customer
  concerns sections with "shadow" appearance anyway, we rather see it
  as a way to make the feature also work with the "cards" appearance
  while visually staying in the realm of the current design.

  This is the approach taken in this commit.

* Placing the heading on a separate card.

  Would work just as well as the previous option. Comes down to a
  matter of taste. Placing the heading outside of the box looked less
  novel than placing two boxes with different widths next to each
  other and was thus preferred.

Based on its similarity with the full position option and the fact
that we currently do not plan to combine the wide setting with other
positions like sticky or left/right, we model the new variant as an
additional position.

Should we ever decide to make content element width configurable
independently of the position setting (e.g. wide and narrow sticky
elements), this decision will not make things worse compared to only
having to deal with the existing full position. Both full and wide
positions could then probably be removed from the editor UI, kept in
the frontend code for backwards compatibility and replaced with a
separate width setting for moving forward.

To prevent the new position option from becoming available for other
content elements, we now need to explicitly define supported positions
for all element types.

REDMINE-18531