Let page types register initializers
We need a place for page types that render embeds to register consent
vendors. We only want to register the vendor if the page type is used
in the entry. We cannot register vendors in `enhance` since page
widgets are created lazily and vendor registration needs to happen
right when the entry loads. Also which vendors to register might
depend on page configuration (for page types that support different
embeds).

We therefore introduce new API to register a page type specific
initializer function that is invoked once per page during entry
initialization.

REDMINE-18003