Prevent mixing in subviewContainer multiple times
In certain cases (e.g., when a Webpacker bundle and an asset pipeline
bundle are both used in a page), `pageflow/ui` may be evaluated
multiple times. When the mixin in included multiple times it leads to
double renders when `appendSubview` is called.

This can be problematic if there is an expensive to render view (in
our case an iframe based theme preview).

REDMINE-19343