Use scrolled entry in seed helper spec
So far the helper did not care about the entry type. To be able to use
entry type specific configuration methods in the helper, we need to be
sure to actually pass a scrolled entry in specs.

Use TestUploadableFile instead of VideoFile when registering file
types in spec to prevent collision with default config of the scrolled
entry type.

REDMINE-19482