Prevent clipping credits box
When the credit box gets long, it is clipped at the bottom on small
screens. This is caused by an overflow hidden container that was
placed there to prevent the display of horizontal scroll bars while
the unmute animation is playing.

Not able to reproduce horizontal overflow in Chrome, FF, Safari and
mobile Safari while animation is playing. Might have been an IE11
issue, which we do not support anymore.

REDMINE-19484