Adds a twitter widget embed as a contentElement
This adds a twitter widget embed as an addition to contentElements
A prescript is loaded to initiate twitter widget, after the widget has
been loaded it uses the twtter.widgets.createTweetEmbed to create
a twitter embed element.
Twitter embed currently takes a tweet-ID, and options to either hide
conversation or hide medias like picture, videos etc.