Pass consistent arguments when transforming theme options
So far, we passed a `PublishedEntry` when actually computing theme
options for a published entry, but only passed an `Entry` when
previewing a theme customization. We now always pass an `Entry`.

To allow having the transformation depend on the actual theme, we
change the signature of the lambda to also take a theme option. This
is safer than always passing a `PublishedEntry` since calling
`PublishedEntry.theme` in the transform leads to a stack overflow.

REDMINE-17625