Make it easier to use FileInputView in outside tests
JST templates in the pageflow package reference the I18n global. In
real life, this works since the global actually exists in the
editor. To make this work in specs, the pageflow package uses a
special JST transform [1]. When importing views that use such
templates from `pageflow/editor` in the `pageflow-scrolled` package,
JST has already been compiled by Rollup, so the transform no longer
applies and the global is undefined.

Since we already moved towards template strings in other places, we
can make things easier and circumvent the problem by inlining the
template.

[1] https://github.com/codevise/pageflow/blob/c8cd95277a6b68ff13df8de229589c42c0c5c4f8/package/config/jest/transformers/jst.js#L6